Jungmo/vote check time

+++ src/main/java/com/example/pj2be/controller/inquirycontroller/InquiryController.java
@@ -4,12 +4,17 @@ import com.example.pj2be.domain.answer.AnswerDTO;
 import com.example.pj2be.domain.inquiry.InquiryDTO;
 import com.example.pj2be.domain.member.MemberDTO;
 import com.example.pj2be.service.inquiryservice.InquiryService;
+import com.example.pj2be.utill.MemberAccess;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Map;
+
+import static com.example.pj2be.utill.MemberAccess.IsLoginMember;
+import static com.example.pj2be.utill.MemberAccess.MemberChecked;
 
 @RestController
 @RequiredArgsConstructor
@@ -29,11 +34,10 @@ public class InquiryController {
     @PostMapping("write")
     public ResponseEntity write(@RequestBody InquiryDTO dto) {
 
-//        로그인 객체 되면 수정
-
         if (!service.validate(dto)) {
             return ResponseEntity.badRequest().build();
         }
+
         service.add(dto);
 
         return ResponseEntity.ok().build();
@@ -46,18 +50,49 @@ public class InquiryController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity delete(@PathVariable Integer id) {
+    public ResponseEntity delete(@PathVariable Integer id,
+                                 @RequestBody InquiryDTO dto) {
 
-        service.delete(id);
-        return ResponseEntity.ok().build();
+        if (IsLoginMember(dto.getLogin_member_id())) {
+            System.out.println("로그인 되어있음: " + dto.getLogin_member_id());
+            if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_member_id()) == 0) {
+                System.out.println("아이디랑 멤버아이디 같음");
+                service.delete(id);
+                System.out.println("수정 완료");
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+
+        if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_category()) == 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        return ResponseEntity.internalServerError().build();
+
     }
 
     @PutMapping("edit")
     public ResponseEntity edit(@RequestBody InquiryDTO dto) {
-        dto.setInquiry_member_id("testadmin");
 
-        service.update(dto);
-        return ResponseEntity.ok().build();
+        if (IsLoginMember(dto.getLogin_member_id())) {
+            System.out.println("로그인 되어있음: " + dto.getLogin_member_id());
+            if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_member_id()) == 0) {
+                System.out.println("아이디랑 멤버아이디 같음");
+                service.update(dto);
+                System.out.println("수정 완료");
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+
+        if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_category()) == 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        return ResponseEntity.internalServerError().build();
     }
 
     @PostMapping("answer")
@@ -67,4 +102,18 @@ public class InquiryController {
         service.answerAdd(dto);
         return ResponseEntity.ok().build();
     }
+
+    @PutMapping("answer/edit")
+    public ResponseEntity answerEdit(@RequestBody AnswerDTO dto) {
+        service.answerEdit(dto);
+        return ResponseEntity.ok().build();
+
+    }
+
+    @DeleteMapping("answer/delete")
+    public ResponseEntity deleteAnswer(@RequestBody AnswerDTO dto) {
+        service.answerDelete(dto);
+        return ResponseEntity.ok().build();
+    }
+
 }
